Add mapper and implement all unused data

+++ app/src/main/java/com/pvz/datadots/data/db/PointEntity.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "points")
 data class PointEntity(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val x: Double,
-    val y: Double
+    val x: Float,
+    val y: Float
 )
